+++ vector/private_vector.h
@@ -6,7 +6,7 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/06 19:31:45 by dbendu            #+#    #+#             */
-/*   Updated: 2020/04/17 15:04:54 by user             ###   ########.fr       */
+/*   Updated: 2020/04/26 21:47:23 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@
 
 typedef struct			s_vector
 {
-	size_t				size;
-	size_t				capacity;
-	size_t				typesize;
+	int					size;
+	int					capacity;
+	int					typesize;
 	void				*data;
 }						t_vector_s;
