+++ vector/vec_reserve.c
@@ -6,7 +6,7 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/06 19:31:43 by dbendu            #+#    #+#             */
-/*   Updated: 2020/03/28 10:27:33 by user             ###   ########.fr       */
+/*   Updated: 2020/04/26 21:50:28 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 #ifdef SAFE_MODE
 
-void			vec_reserve(t_vector vector, size_t elems)
+void			vec_reserve(t_vector vector, int elems)
 {
 	t_vector_s	*vec;
-	size_t		total_cells;
+	int			total_cells;
 
 	if (!vector)
 		ft_error("invalid param \"vecptr\": cannot be NULL)",
@@ -34,10 +34,10 @@ void			vec_reserve(t_vector vector, size_t elems)
 
 #else
 
-void			vec_reserve(t_vector vector, size_t elems)
+void			vec_reserve(t_vector vector, int elems)
 {
 	t_vector_s	*vec;
-	size_t		total_cells;
+	int			total_cells;
 
 	vec = *(void**)vector - sizeof(t_vector_s);
 	total_cells = vec->capacity + vec->size;
